pyproject.toml
[project]
name = "morf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django>=5.2.3",
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.16.0",
]
